[WiP] OSprofiler with OpenTracing compatible
This is the very first PoC of OpenTracing compatible in OSprofiler.

In this patch, I use Uber Jaeger as a tracer.
Trace/span from OSprofiler will be sent to Jaeger.
We can use Jaeger UI to view trace.

I used Jaeger all-in-one and jaeger-client to demonstrate this PoC.

A sample image of trace in Jaeger:
https://tovin07.github.io/opentracing/jaeger-openstack-image-list.png

Reference:
- OpenTracing: http://opentracing.io/
- Uber Jager: http://jaeger.readthedocs.io/en/latest/

Additional requirements (for this PoC):
- Docker engine to run jaeger all-in-one
- `pip install jaeger-client` to instrument OSprofiler

Change-Id: I8f75af487f50d5e67a1321a4c951d2a0d85733a1
Implements: blueprint opentracing-compatible